Struct slc::room_controller::RoomController

pub struct RoomController {
    pub room_data: RoomData,
    // some fields omitted
}

Fields
room_data: RoomData
Implementations

impl RoomController

pub fn new(filepath: &str) -> RoomController

Creates a RoomController by first creating and consuming a RoomData struct. Also precalculates the angle, direction, and position of each LED, relative to the view position.

pub fn set(&mut self, index: usize, color: Color)

Sets the color of a given led

pub fn set_all(&mut self, color: Color)

Sets the color of all leds in the room

pub fn set_at_view_dir(
&mut self,
dir: Vector2D,
color: Color,
enable_smoothing: bool
)

Sets the color of the pixel in a given direction, relative to the view.

pub fn set_at_view_angle(
&mut self,
angle: f32,
color: Color,
enable_smoothing: bool
)

Sets the color of the pixel at a given angle, relative to the view.

pub fn set_at_room_angle(
&mut self,
angle: f32,
color: Color,
enable_smoothing: bool
)

Sets the color of the pixel at a given angle, relative to the room.

pub fn get_led_at_room_dir(&self, dir: Vector2D) -> Option<(usize, f32)>

Casts a ray in the given direction, in room coordinate space, from the camera’s position. If it hits a wall, the id of the led closest to that wall position will be returned, as well as the “Occupancy” of that led, where 1.0 means the ray lands directly on the LED, and 0.5 means the ray is halfway between that led and the next one up.

pub fn get_color_at_room_dir(&self, dir: Vector2D) -> Color

Returns the color of the led at the given room-space direction. If no led exists in that direction, black is returned.

pub fn set_at_room_dir(
&mut self,
dir: Vector2D,
color: Color,
enable_smoothing: bool
)

Uses get_led_at_room_dir() to color an led at a given room-space direction.

pub fn map_angle_to_color(&mut self, map: &dyn Fn(f32) -> Color)

Allows the user to pass in a Color-returning function to calculate the color of each led, given its angle.

pub fn map_angle_to_color_clamped(
&mut self,
map: &dyn Fn(f32) -> Color,
min_angle: f32,
max_angle: f32
)

Allows the user to pass in a Color-returning function to calculate the color of each led within a range, given its angle.

pub fn map_dir_to_color(&mut self, map: &dyn Fn(Vector2D) -> Color)

Allows the user to pass in a Color-returning function to calculate the color of each led, given its direction.

pub fn map_dir_to_color_clamped(
&mut self,
map: &dyn Fn(Vector2D) -> Color,
min_angle: f32,
max_angle: f32
)

Allows the user to pass in a Color-returning function to calculate the color of each led within an angle range, given its direction.