Struct slc::room_data::RoomData

pub struct RoomData {
    pub brightness: u8,
    // some fields omitted
}

Fields
brightness: u8

Brightness of the leds on a 0-255 scale.

Implementations

impl RoomData

pub fn new_from_file(filepath: &str) -> Self

constructs a room from a .rcfg file.

pub fn density(&self) -> f32

read-only access to the density field.

pub fn view_pos(&self) -> Point

read-only access to the view_pos field.

pub fn view_rot(&self) -> f32

read-only access to the view_rot field.

pub fn strips(&self) -> &Vec<Strip>

read-only access to the srips field.

pub fn leds(&self) -> &Vec<Color>

read-only access to the leds field.

pub fn view_dir_to_room_dir(&self, view_dir: Vector2D) -> Vector2D

Transforms a direction in view space to room space.

pub fn view_angle_to_room_angle(&self, view_angle: f32) -> f32

Transforms an angle in view space to room space.

pub fn get_pos_at_t(&self, t: f32) -> Point

Interpolates down the chain of strips to return the point at t.

pub fn set_led(&mut self, index: usize, color: Color)

Sets the led at index to the given color.